feat: use index barrel for db mappings
