+++ application/app/lib/db-mapping/index.ts
@@ -1,0 +1,5 @@
+export * from './communes';
+export * from './departements';
+export * from './etablissements';
+export * from './regions';
+export * from './searchView';
